Deep_Web_Research_Agentic_System/requirements.lock.txt
ag2==0.9.6
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asttokens==3.0.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.9.6
autogen-agentchat==0.6.4
autogen-core==0.6.4
babel==2.17.0
beautifulsoup4==4.13.4
bibtexparser==1.4.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
ddgs==9.2.3
debugpy==1.8.14
decorator==5.2.1
deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docutils==0.21.2
executing==2.2.0
fake-useragent==2.2.0
fastapi==0.116.1
free-proxy==1.1.3
frozenlist==1.7.0
greenlet==3.2.3
groq==0.30.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
imagesize==1.4.1
importlib-metadata==8.7.0
ipykernel==6.29.5
ipython==9.4.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jupyter-client==8.6.3
jupyter-core==5.8.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-groq==0.3.6
langchain-ollama==0.3.4
langchain-tavily==0.2.10
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langgraph==0.5.3
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.73
langsmith==0.4.8
lxml==6.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.6.3
mypy-extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.1
ollama==0.5.1
opentelemetry-api==1.35.0
orjson==3.11.0
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==25.0
parso==0.8.4
pillow==11.3.0
pip==25.1.1
platformdirs==4.3.8
primp==0.15.0
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pure-eval==0.2.3
pyautogen==0.10.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyparsing==3.2.3
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywin32==311
pyyaml==6.0.2
pyzmq==27.0.0
redis==6.2.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
roman-numerals-py==3.1.0
scholarly==1.7.11
selenium==4.34.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.7
sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlalchemy==2.0.41
stack-data==0.6.3
starlette==0.47.1
tenacity==9.1.2
termcolor==3.1.0
tiktoken==0.9.0
tornado==6.5.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20250708
types-requests==2.32.4.20250611
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.5.0
uv==0.7.21
uvicorn==0.35.0
wcwidth==0.2.13
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
